Improved information architecture of Arrival prediction screen.

+++ app/src/main/java/mx/x10/filipebezerra/horariosrmtcgoiania/api/model/BusTravelModel.java
@@ -23,5 +23,5 @@ public class BusTravelModel {
     public String expectedArrivalTime;
 
     @SerializedName("PrevisaoChegada")
-    public int minutesToArrive;
+    public Long minutesToArrive;
 }
